BLPOP changes

+++ src/main/java/server/BlockedClient.java
@@ -8,10 +8,10 @@ public class BlockedClient {
     private final List<String> keys;
     private final long timeoutAt;
     private final long blockedAt;
-    private final int timeout;
+    private final double timeout;
 
 
-    public BlockedClient(SocketChannel channel, List<String> keys, int timeout){
+    public BlockedClient(SocketChannel channel, List<String> keys, double timeout){
         this.channel = channel;
         this.keys = keys;
         this.blockedAt = System.currentTimeMillis();
@@ -19,7 +19,7 @@ public class BlockedClient {
         if(timeout == 0){
             this.timeoutAt = Long.MAX_VALUE;
         }else{
-            this.timeoutAt = System.currentTimeMillis()+(timeout*1000L);
+            this.timeoutAt = (long) (System.currentTimeMillis()+(timeout*1000L));
         }
         this.timeout = timeout;
     }
